requirements.txt
click
click_config_file
Jinja2
Mastodon.py
bs4
